Mv proxy tests to a separate module
